backend/requirements.txt
clickhouse-connect>=0.7.0
kafka-python>=2.0.2
fastavro>=1.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
prometheus_client>=0.16.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.39b0
opentelemetry-exporter-otlp>=1.20.0
python-json-logger>=2.0.7
